requirements.txt
Flask
Flask-CORS
python-dotenv
ccxt
pandas
ta
requests